refactor: 테스트 컨테이너 설정 패키지 분리
