+++ backend/src/test/java/com/mapmory/backend/support/MySqlTestContainerConfig.java
@@ -1,4 +1,4 @@
-package com.mapmory.backend;
+package com.mapmory.backend.support;
 
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
